+++ kvstore/badger/utils.go
@@ -1,0 +1,27 @@
+package badger
+
+import "os"
+
+// Returns whether the given file or directory exists.
+func exists(path string) (bool, error) {
+	_, err := os.Stat(path)
+	if err == nil {
+		return true, nil
+	}
+	if os.IsNotExist(err) {
+		return false, nil
+	}
+	return false, err
+}
+
+func checkDir(dir string) error {
+	exists, err := exists(dir)
+	if err != nil {
+		return err
+	}
+
+	if !exists {
+		return os.Mkdir(dir, 0700)
+	}
+	return nil
+}
